python/pyproject.toml
[tool.poetry]
name = "python-demo"
version = "0.1.0"
description = ""
authors = ["Victor França Lopes <18534956+vflopes@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "petstore"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-lambda-typing = "^2.17.1"
grpcio = "^1.56.2"
protobuf = "^4.23.4"
protomesh-py = { git = "https://github.com/protomesh/protomesh-py.git", rev = "9aa18b70e60728f2dbce3342e81bfcdb06479463" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
